Modifies sign portal detections to allow arbitrarily large portals
Modifies sign portal detection to allow a portal which has any size (width and height). The sign must be placed in the bottom corner of a portal (i.e. if you build a normal portal, knock out the bottom corner block and put a sign on the obsidian inside that corner)